Improve recipes internally Add exchange recipes for vanilla cord and tinkers' string cord Yoyo entities now use entity data Yoyo constructor takes a hand now, simplifies code

+++ src/main/java/com/jozufozu/yoyos/client/NetworkHandlers.java
@@ -48,7 +48,7 @@ public class NetworkHandlers
     {
         if (event.phase != TickEvent.Phase.START) return;
 
-        int reel = 0;
+        byte reel = 0;
 
         Minecraft mc = Minecraft.getMinecraft();
         EntityYoyo yoyo = EntityYoyo.CASTERS.get(mc.player);
